refactor: eliminate code duplication and unnecessary allocations
Ultra-deep review identified three significant optimization opportunities:

**1. Extract Size Parsing (Eliminates 45 Lines of Duplication)**
- Created shared `util::parse_size()` function
- Replaced duplicate implementations in main.rs and typechecker.rs
- Single source of truth for size parsing with consistent error messages

**2. Remove Unnecessary Vec Allocations (6 instances)**
File: `src/eval/structured.rs`
- Removed redundant `.to_vec()` calls that copied Vec<&T> → Vec<&T>
- Eliminated intermediate `refs` variables
- Pass `&values` directly to comparison functions (accepts &[&T])
- Removes 6 unnecessary allocations per structured query evaluation

Before:
```rust
let values = navigate_yaml(doc, path);
let refs: Vec<&yaml_rust::Yaml> = values.to_vec();  // Wasteful!
if compare_yaml_values(&refs, *operator, value, raw_string)
```

After:
```rust
let values = navigate_yaml(doc, path);
if compare_yaml_values(&values, *operator, value, raw_string)
```

**3. Make util Module Public**
- Export util module in lib.rs for access from main.rs

**Impact**:
- **~57 lines removed**
- **Eliminates 6 Vec allocations** per file evaluation with structured selectors
- **Single source of truth** for size parsing
- All tests pass, no clippy warnings
